add: Parameterization of the entire macaroni noodle
Note that there are two inverted tets in the final parameterization,
but these are sliver tets in the original mesh.  I need to find a
way to edit the mesh for those tets so that I can make this a passing
test without having to redo the parameterization of the surface on a
new mesh.