fix: Don't use tolerancing to avoid tracing to edges adjacent to start vertex on boundary
The removal at Tracing.cpp -372,9 +415,6 is replaced by refactoring a bunch
of stuff to not trace to edges adjacent to a start vertex, if the start is
a vertex.  The tolerancing being removed was the wrong solution, and was
breaking things elsewhere.

This only fixes the boundary.  A similar refactor for interior tracing will
be submitted in a separate fix.